delete ok

+++ controllers/dashboard/vehicles/deleteVehicles-ctrl.php
@@ -7,20 +7,30 @@ session_start();
 try {
     $id_vehicle = intval(filter_input(INPUT_GET, 'id_vehicle', FILTER_SANITIZE_NUMBER_INT)); // récupérer la donnée tout en la nettoyant, ne pas utiliser $_GET. intval permet de retourner un entier dans tous les cas, 1 au lieu de "1"
 
-    $delete = Vehicle::delete($id_vehicle);
+    $picture = Vehicle::getPictureUpdate($id_vehicle);
 
-    if ($delete) {
+    $fileName = __DIR__ . '/../../../public/uploads/users/' . $picture->picture;
+    // var_dump($fileName);
 
-        
-        $msg = 'La donnée a bien été supprimée !';
-        $_SESSION['msg'] = $msg;
-        
+    $deletePic = unlink($fileName); // supprimer l'image du disque local
+    // var_dump($deletePic);
 
-        header('Location: /controllers/dashboard/vehicles/listVehicles-ctrl.php');
-        die;
-    } else {
-        $msg = 'La donnée n\'a pas été supprimée !';
+    if ($deletePic) {
+        $delete = Vehicle::delete($id_vehicle);
+
+        if ($delete) {
+
+            $msg = 'La donnée a bien été supprimée !';
+            $_SESSION['msg'] = $msg;
+            
+    
+            header('Location: /controllers/dashboard/vehicles/listVehicles-ctrl.php');
+            die;
+        } else {
+            $msg = 'La donnée n\'a pas été supprimée !';
+        }
     }
+
 
 
 } catch (\Throwable $th) {
